requirements.txt
django
django-extensions
djangorestframework
djangorestframework-simplejwt
pillow
django-pagination